# specs/tsp-output/@typespec/openapi3/generated/openapi.yaml
openapi: 3.0.0
info:
  title: (title)
  description: Comprehensive API for managing beauty salon operations including salons, staff, services, customers, reservations, bookings, and reviews.
  version: 0.0.0
...
paths:
  /admin/auth/trusted-ip/{userId}:
    post:
      operationId: AdminAuthOperations_addTrustedIp
      description: Add trusted IP address for a user
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.AuthSuccessResponse'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
      tags:
        - Admin - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ipAddress:
                  type: string
              required:
                - ipAddress
    delete:
      operationId: AdminAuthOperations_removeTrustedIp
      description: Remove trusted IP address for a user
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: ipAddress
          in: query
          required: true
          schema:
            type: string
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.AuthSuccessResponse'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
      tags:
        - Admin - Authentication
...
  /auth/verify-email/confirm:
    post:
      operationId: AuthOperations_confirmEmailVerification
      description: Confirm email verification
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.AuthSuccessResponse'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Models.EmailVerificationRequest'
  /auth/verify-email/send:
    post:
      operationId: AuthOperations_sendEmailVerification
      description: Send email verification
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.AuthSuccessResponse'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
        '429':
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
      tags:
        - Authentication
...
components:
...
  schemas:
...
    Models.AuthSuccessResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Success message
      description: Generic auth success response
...
    Models.EmailVerificationRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: Email verification token
      description: Email verification request
    Models.Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: {}
...
servers:
  - url: https://api.beauty-salon.com
    description: Beauty Salon API
    variables: {}